+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to BangoCat-mac will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased] - 2025-07-28
+
+### Modified
+
+- Add public methods for checking updates and opening preferences in AppDelegate; update context menu in CatView to include new options
+
+
+
 ## [1.4.3] - 2025-07-28
 
 ### Modified